refactor: 소모품 consumable > item 명칭 변경
