+++ android/feature/home/src/main/java/com/obrit/feature/home/screen/homeSection/ItemOrbitSection.kt
@@ -63,7 +63,7 @@ import kotlin.math.sqrt
  * @param icons 볼 안에 표시할 소모품 아이콘 목록. 변경 시 물리 상태가 초기화된다.
  */
 @Composable
-internal fun ConsumableOrbitSection(
+internal fun ItemOrbitSection(
     icons: List<ConsumableIcon>,
     normalRatio: Float,
     warningRatio: Float,
@@ -800,9 +800,9 @@ private val DetergentHeight = 54.dp
 
 @Preview(showBackground = true, backgroundColor = 0xFF1D1B20)
 @Composable
-private fun ConsumableOrbitSectionPreview() {
+private fun ItemOrbitSectionPreview() {
     OBRitTheme {
-        ConsumableOrbitSection(
+        ItemOrbitSection(
             icons =
                 listOf(
                     ConsumableIcon(R.drawable.ic_towel, TowelWidth, TowelHeight),
