+++ android/feature/home/src/main/java/com/obrit/feature/home/screen/homeSection/QuickItemSection.kt
@@ -1,290 +1,185 @@
-@file:Suppress("TooManyFunctions")
+@file:Suppress("TooManyFunctions") @file:JvmName("QuickItemSectionKt")
 
 package com.obrit.feature.home.screen.homeSection
 
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material3.Icon
-import androidx.compose.material3.Text
+import androidx.compose.foundation.lazy.LazyRow
+import androidx.compose.foundation.lazy.items
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.draw.drawBehind
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.drawscope.drawIntoCanvas
-import androidx.compose.ui.graphics.nativeCanvas
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.window.Dialog
-import androidx.compose.ui.window.DialogProperties
-import androidx.core.content.ContextCompat
-import coil.compose.AsyncImage
-import com.obrit.android.core.designsystem.R
-import com.obrit.android.core.designsystem.component.bottomsheet.OBRitBottomSheet
-import com.obrit.android.core.designsystem.component.button.OBRitButtonDefaults
-import com.obrit.android.core.designsystem.component.button.OBRitLargeFilledButton
+import com.obrit.android.core.designsystem.component.card.OBRitCardGrid
 import com.obrit.android.core.designsystem.component.card.OBRitCardLevel
-import com.obrit.android.core.designsystem.component.card.OBRitCardList
-import com.obrit.android.core.designsystem.theme.LocalOBRitColor
-import com.obrit.android.core.designsystem.theme.LocalOBRitTypography
+import com.obrit.android.core.designsystem.component.chip.OBRitChip
 import com.obrit.android.core.designsystem.theme.OBRitTheme
-import com.obrit.feature.home.viewmodel.ConsumableListSortOrder
-import com.obrit.obrit.shared.designsystem.tokens.atom.radius.AtomRadius
 import com.obrit.obrit.shared.designsystem.tokens.atom.spacing.AtomSpacing
-import com.obrit.obrit.shared.model.home.HomeItemCard
-import kotlin.math.roundToInt
-
-private const val LIST_PREVIEW_COUNT = 3
+import com.obrit.obrit.shared.model.ReplacementDate
+import com.obrit.obrit.shared.model.home.HomeBucketGroup
+import com.obrit.obrit.shared.model.home.HomeBucketItem
+import com.obrit.obrit.shared.model.home.HomeBucketType
+import com.obrit.obrit.shared.model.home.HomeStatusLevel
+import java.time.LocalDate
+import java.time.temporal.ChronoUnit
 
 @Composable
 internal fun QuickItemSection(
-    items: List<HomeItemCard>,
-    sortOrder: ConsumableListSortOrder,
-    onSortOrderChange: (ConsumableListSortOrder) -> Unit,
-    onMoreClick: () -> Unit,
+    buckets: List<HomeBucketGroup>,
     modifier: Modifier = Modifier,
 ) {
-    val typography = LocalOBRitTypography.current
-    var showSortSheet by remember { mutableStateOf(false) }
-
-    Column(
-        modifier = modifier.padding(horizontal = AtomSpacing.S5.dp, vertical = AtomSpacing.S5.dp),
-        verticalArrangement = Arrangement.spacedBy(AtomSpacing.S3.dp),
-    ) {
-        Row(verticalAlignment = Alignment.CenterVertically) {
-            SortDropdown(
-                sortOrder = sortOrder,
-                onClick = { showSortSheet = true },
-            )
-
-            Text(
-                text = "미리보기",
-                style = typography.xl,
-                color = LocalOBRitColor.current.common00,
-                modifier = Modifier.padding(horizontal = 12.dp, vertical = 7.dp),
-            )
+    if (buckets.isEmpty()) return
+    val hasDangerItems = buckets.find { it.bucket == HomeBucketType.DANGER }?.items?.isNotEmpty() == true
+    var selectedType by remember { mutableStateOf(if (hasDangerItems) buckets.first().bucket else buckets.last().bucket) }
+    val selectedItems =
+        remember(buckets, selectedType) {
+            buckets.find { it.bucket == selectedType }?.items ?: emptyList()
         }
 
-        QuickItemList(items = items.take(LIST_PREVIEW_COUNT))
-        OBRitLargeFilledButton(
-            text = "더보기",
-            onClick = onMoreClick,
-            colors = OBRitButtonDefaults.defaultButtonColors(),
-            modifier = Modifier.fillMaxWidth(),
+    Column(
+        modifier = modifier.padding(vertical = AtomSpacing.S5.dp),
+        verticalArrangement = Arrangement.spacedBy(AtomSpacing.S3.dp),
+    ) {
+        BucketFilterChipRow(
+            buckets = buckets,
+            selectedType = selectedType,
+            onTypeSelect = { selectedType = it },
         )
-    }
-
-    if (showSortSheet) {
-        SortBottomSheet(
-            sortOrder = sortOrder,
-            onSortOrderChange = {
-                onSortOrderChange(it)
-                showSortSheet = false
-            },
-            onDismiss = { showSortSheet = false },
-        )
+        BucketCardRow(items = selectedItems)
     }
 }
 
 @Composable
-private fun SortDropdown(
-    sortOrder: ConsumableListSortOrder,
+private fun BucketFilterChipRow(
+    buckets: List<HomeBucketGroup>,
+    selectedType: HomeBucketType,
+    onTypeSelect: (HomeBucketType) -> Unit,
     modifier: Modifier = Modifier,
-    onClick: (() -> Unit)? = null,
 ) {
-    val typography = LocalOBRitTypography.current
-    val colors = LocalOBRitColor.current
-    val context = LocalContext.current
-    val backgroundDrawable =
-        ContextCompat.getDrawable(context, R.drawable.ic_filter_chip_background)
     Row(
         modifier =
             modifier
-                .clip(RoundedCornerShape(AtomRadius.Small.dp))
-                .drawBehind {
-                    backgroundDrawable?.let { drawable ->
-                        drawable.setBounds(0, 0, size.width.roundToInt(), size.height.roundToInt())
-                        drawIntoCanvas { drawable.draw(it.nativeCanvas) }
-                    }
-                }.then(if (onClick != null) Modifier.clickable(onClick = onClick) else Modifier)
-                .padding(horizontal = AtomSpacing.S3.dp, vertical = AtomSpacing.S2.dp),
-        verticalAlignment = Alignment.CenterVertically,
-        horizontalArrangement = Arrangement.spacedBy(AtomSpacing.S1.dp),
+                .fillMaxWidth()
+                .padding(horizontal = AtomSpacing.S5.dp),
+        horizontalArrangement = Arrangement.spacedBy(AtomSpacing.S2.dp),
     ) {
-        Text(
-            text = sortOrder.displayName,
-            style = typography.base.copy(fontWeight = FontWeight.SemiBold),
-            color = colors.common00,
-            maxLines = 1,
-        )
-
-        Icon(
-            painter = painterResource(id = R.drawable.ic_dropdown_chevron_down),
-            contentDescription = null,
-            tint = Color.Unspecified,
-            modifier = Modifier.size(AtomSpacing.S4.dp),
-        )
-    }
-}
-
-@Composable
-private fun SortBottomSheet(
-    sortOrder: ConsumableListSortOrder,
-    onSortOrderChange: (ConsumableListSortOrder) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    Dialog(
-        onDismissRequest = onDismiss,
-        properties = DialogProperties(usePlatformDefaultWidth = false),
-    ) {
-        SheetContainer(onDismiss = onDismiss) {
-            SortBottomSheetContent(
-                sortOrder = sortOrder,
-                onSortOrderChange = onSortOrderChange,
+        buckets.filter { it.count > 0 }.forEach { group ->
+            OBRitChip(
+                text = group.bucket.displayName,
+                number = group.count,
+                selected = group.bucket == selectedType,
+                onClick = { onTypeSelect(group.bucket) },
             )
         }
     }
 }
 
 @Composable
-private fun SheetContainer(
-    onDismiss: () -> Unit,
-    content: @Composable ColumnScope.() -> Unit,
-) {
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .background(SORT_SHEET_SCRIM_COLOR)
-                .clickable(onClick = onDismiss),
-        contentAlignment = Alignment.BottomCenter,
-    ) {
-        OBRitBottomSheet(
-            modifier = Modifier.clickable(onClick = {}),
-            content = content,
-        )
-    }
-}
-
-@Composable
-private fun SortBottomSheetContent(
-    sortOrder: ConsumableListSortOrder,
-    onSortOrderChange: (ConsumableListSortOrder) -> Unit,
+private fun BucketCardRow(
+    items: List<HomeBucketItem>,
     modifier: Modifier = Modifier,
 ) {
-    val typography = LocalOBRitTypography.current
-    val colors = LocalOBRitColor.current
-    Column(
-        modifier =
-            modifier
-                .fillMaxWidth()
-                .padding(vertical = AtomSpacing.S4.dp, horizontal = AtomSpacing.S5.dp),
+    LazyRow(
+        modifier = modifier.fillMaxWidth(),
+        contentPadding = PaddingValues(horizontal = AtomSpacing.S5.dp),
+        horizontalArrangement = Arrangement.spacedBy(AtomSpacing.S3.dp),
     ) {
-        ConsumableListSortOrder.entries.forEach { order ->
-            Row(
-                modifier =
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(vertical = AtomSpacing.S4.dp)
-                        .clickable { onSortOrderChange(order) },
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                Text(
-                    text = order.displayName,
-                    style = typography.xl.copy(fontWeight = FontWeight.Bold),
-                    color = colors.common00,
-                    modifier = Modifier.weight(1f),
-                )
-                if (sortOrder == order) {
-                    Icon(
-                        painter = painterResource(id = R.drawable.ic_sort_selected),
-                        contentDescription = null,
-                        tint = Color.Unspecified,
-                        modifier = Modifier.size(AtomSpacing.S6.dp),
-                    )
-                }
-            }
+        items(items) { item ->
+            BucketCard(item = item)
         }
     }
 }
 
 @Composable
-private fun QuickItemList(
-    items: List<HomeItemCard>,
-    modifier: Modifier = Modifier,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(AtomSpacing.S2.dp),
-    ) {
-        items.forEach { item ->
-            QuickItemListItem(item = item)
-        }
-    }
-}
-
-@Composable
-internal fun QuickItemListItem(item: HomeItemCard) {
-    OBRitCardList(
-        level = homeItemCardLevel(item),
+private fun BucketCard(item: HomeBucketItem) {
+    OBRitCardGrid(
+        level = itemCardLevel(item.status),
         title = item.name,
-        daysInUseLabel = "${item.daysInUse}일",
-        replaceLabel = item.replacementDday,
-        sparesLabel = "여분 ${item.spareQuantity}개",
-        image = {
-            if (item.iconUrl.isNotEmpty()) {
-                AsyncImage(
-                    model = item.iconUrl,
-                    contentDescription = item.name,
-                )
-            }
-        },
+        stockCount = item.spareQuantity,
+        daysLabel = daysLabel(item.nextReplacementDate),
     )
 }
 
-@Suppress("MagicNumber")
-private fun homeItemCardLevel(item: HomeItemCard): OBRitCardLevel {
-    val hasSpare = item.spareQuantity > 0
-    return if (hasSpare) OBRitCardLevel.L1 else OBRitCardLevel.L2
+private val HomeBucketType.displayName: String
+    get() =
+        when (this) {
+            HomeBucketType.DANGER -> "위험"
+            HomeBucketType.WARNING -> "경고"
+            HomeBucketType.UNKNOWN -> ""
+        }
+
+private fun itemCardLevel(status: HomeStatusLevel): OBRitCardLevel =
+    when (status) {
+        HomeStatusLevel.DANGER -> OBRitCardLevel.L1
+        HomeStatusLevel.WARNING -> OBRitCardLevel.L4
+        else -> OBRitCardLevel.L6
+    }
+
+private fun daysLabel(replacementDate: ReplacementDate?): String {
+    replacementDate ?: return "-"
+    val days =
+        ChronoUnit.DAYS
+            .between(LocalDate.now(), LocalDate.parse(replacementDate.value))
+            .toInt()
+    return when {
+        days == 0 -> "D-day"
+        days > 0 -> "D-$days"
+        else -> "D+${-days}"
+    }
 }
 
 @Suppress("MagicNumber")
-private val SORT_SHEET_SCRIM_COLOR = Color(0x99000000)
-
-@Suppress("MagicNumber")
-@Preview(showBackground = true, backgroundColor = 0xFF1D1B20, widthDp = 412)
-@Composable
-private fun QuickItem() {
-    OBRitTheme {
-        QuickItemSection(
+private val previewBuckets =
+    listOf(
+        HomeBucketGroup(
+            bucket = HomeBucketType.DANGER,
+            count = 2,
             items =
                 listOf(
-                    HomeItemCard(1L, "면도기", 30, "교체 D+7", 0, ""),
-                    HomeItemCard(2L, "칫솔", 27, "교체 D-day", 1, ""),
-                    HomeItemCard(3L, "샴푸", 22, "교체 D-2", 0, ""),
-                    HomeItemCard(4L, "필터", 10, "교체 D-10", 3, ""),
+                    HomeBucketItem(
+                        itemId = 1,
+                        name = "면도기",
+                        spareQuantity = 0,
+                        nextReplacementDate = ReplacementDate("2026-05-23"),
+                        status = HomeStatusLevel.DANGER,
+                    ),
+                    HomeBucketItem(
+                        itemId = 2,
+                        name = "칫솔",
+                        spareQuantity = 1,
+                        nextReplacementDate = ReplacementDate("2026-05-26"),
+                        status = HomeStatusLevel.DANGER,
+                    ),
                 ),
-            sortOrder = ConsumableListSortOrder.REPLACE_IMMINENT,
-            onSortOrderChange = {},
-            onMoreClick = {},
-        )
+        ),
+        HomeBucketGroup(
+            bucket = HomeBucketType.WARNING,
+            count = 1,
+            items =
+                listOf(
+                    HomeBucketItem(
+                        itemId = 3,
+                        name = "필터",
+                        spareQuantity = 3,
+                        nextReplacementDate = ReplacementDate("2026-05-26"),
+                        status = HomeStatusLevel.WARNING,
+                    ),
+                ),
+        ),
+    )
+
+@Preview(showBackground = true, backgroundColor = 0xFF1D1B20)
+@Composable
+private fun QuickItemSectionPreview() {
+    OBRitTheme {
+        QuickItemSection(buckets = previewBuckets)
     }
 }
